requirements.txt
asn1crypto==1.4.0
attrs==21.2.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
endesive==2.0.9
fpdf2==2.4.1
idna==2.10
lxml==4.6.3
oscrypto==1.2.1
paramiko==2.7.2
pdfrw==0.4
Pillow==8.2.0
pycparser==2.20
PyKCS11==1.5.10
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pypng==0.0.20
PyQRCode==1.2.1
python-decouple==3.4
python-pkcs11==0.7.0
pytz==2021.1
qrcode==6.1
reportlab==3.5.67
requests==2.25.1
six==1.16.0
urllib3==1.26.5
watchdog==2.1.2